requirements.txt
arduino-iot-cloud
arduino-iot-client
matplotlib
oauthlib
requests
requests_oauthlib
pyyaml